Refactor: Create TimerData.cpp
In this PR we're creating TimerData.cpp and moving most implementations
from TimerData.h because that class is too big to only have an .h file.
